Extract connection creation assistance to dedicated module
