+++ src/lsp/protocol.ts
@@ -561,7 +561,7 @@ export interface SetIssueStatusParams {
 
 export interface AssistCreatingConnectionParams {
   isSonarCloud: boolean;
-  serverUrl: string;
+  serverUrlOrOrganisationKey: string;
   token: string;
 }
